move function into separate helper object
